requirements.txt
2to3==1.0
asgiref==3.5.2
attrs==21.4.0
auth==0.5.3
blinker==1.4
boto==2.49.0
boto3==1.24.2
botocore==1.27.2
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
contextlib2==21.6.0
cryptography==37.0.2
declxml==0.9.1
defusedxml==0.7.1
dj-database-url==0.5.0
Django==4.0.5
django-allauth==0.50.0
django-cors-headers==3.12.0
django-crontab==0.7.1
django-environ==0.8.1
django-heroku==0.3.1
django-rest-auth==0.9.5
django-rest-framework==0.1.0
django-storages==1.12.3
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
dnspython==2.2.1
eventlet==0.33.0
falcon==3.1.0
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
image==1.5.33
imutils==0.5.4
install==1.3.5
jmespath==1.0.0
jsonschema==4.6.0
mongoengine==0.24.1
numpy==1.22.4
oauthlib==3.2.0
opencv-python==4.5.5.64
packaging==21.3
Pillow==9.1.1
psycopg2==2.9.3
psycopg2-binary==2.9.3
pycparser==2.21
PyJWT==1.7.1
pymongo==4.1.1
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.1
requests==2.27.1
requests-oauthlib==1.3.1
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.26.9
Werkzeug==2.1.2
whitenoise==6.1.0
